Create an object of class 'SansSouciStruct' with a dyadic structure

Create an object of class 'SansSouciStruct' with a dyadic structure

Usage

SansSouciDyadic(
  m,
  leaf_size = NULL,
  height = NULL,
  direction = c("bottom-up", "top-down"),
  ...
)

Arguments

m

A numeric value, the number of hypotheses

leaf_size

An integer value, the number of hypotheses in each leaf. Defaults to NULL.

height

An integer value, the desired maximal height of the tree. Defaults to NULL.

direction

A character value, the direction used for building the tree. Must be one of "bottom-up" or "top-down"

...

Further arguments to be passed to 'SansSouciStruct'

Value

An object of class 'SansSouciStruct'

Details

If both 'leaf_size' and 'height' are NULL then a binary tree with maximal height given the total number of hypotheses is built. If both 'leaf_size' and 'height' are non NULL, then 'height' is ignored and a tree is built with the desired leaf size.

References

Durand, G., Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc false positive control for structured hypotheses. Scandinavian Journal of Statistics, 47(4), 1114-1148.

See also

Examples

s <- 100
q <- 7
m <- s*2^q

obj <- SansSouciDyadic(m, leaf_size = s, direction = "top-down")
obj <- SansSouciDyadic(m, height = 8, direction = "top-down")