kFWERThresholdFamily

calibration of the 'balanced' threshold family of BNR

kFWERThresholdFamily(mat, kMax = nrow(mat), Rcpp = TRUE, verbose = FALSE)

Arguments

mat
A \(m\) x \(B\) matrix of Monte-Carlo samples of test statistics under the null hypothesis.

m

is the number of tested hypotheses

B

is the number of Monte-Carlo samples

kMax
For simultaneous control of (\(k\)-FWER for all \(k \leq k[max]\)).

Rcpp
If TRUE (the default), some costly operations (sorting) are performed in C++.

verbose
If TRUE, print results of intermediate calculations. Defaults to FALSE.

Value

A function \(f:alpha \mapsto f(alpha)\) such that for each \(alpha\), \(f(alpha)\) is a candidate threshold family for controlling the joint FWER.

Details

This family \((t_k)\) of thresholds is calibrated in such a way that for each \(k\leq kMax\), \((t_k)\) controls the (marginal) k-FWER

The \(m\) x \(B\) matrix resulting of sorting the argument mat by row and then by columns is returned as the attribute 'Q' of the return value.

Examples

m <- 1023
B <- 1e3

flavor <- c("independent", "equi-correlated", "3-factor model")[2]
rho <- 0.2
mat <- simulateGaussianNullsFromFactorModel(m, B, flavor=flavor, rho=rho)

sk <- kFWERThresholdFamily(mat)

thr <- sk(0.2)