partialColSortDesc

partial sorting of the columns of a matrix by descending order

partialColSortDesc(mat, k = nrow(mat), Rcpp = TRUE)

Arguments

mat
a m x B numeric matrix

k
an integer value between 1 and m

Rcpp
sorting is performed in C++ if TRUE (the default), and using base::sort if FALSE.

Value

a k x B numeric matrix, whose b-th column corresponds to the k largest values of the b-th column of mat, sorted decreasingly.

Examples

A <- matrix(rnorm(15), 5, 3);
print(A)

#>            [,1]       [,2]      [,3]
#> [1,] -1.3556258  0.9137939 0.8518638
#> [2,] -1.0838822 -0.1579833 0.2272061
#> [3,]  0.2050227  0.6496097 1.4690375
#> [4,]  1.6233358  0.2083355 0.7919008
#> [5,]  2.1451688  0.7536676 0.9914432

BR <- partialColSortDesc(A, 2, Rcpp=FALSE);
B <- partialColSortDesc(A, 2, Rcpp=TRUE);
stopifnot(identical(B,BR));

if (require("rbenchmark")) {
    A <- matrix(rnorm(1e6), 10000, 1000);
    kmax <- 100;
    benchmark(partialColSortDesc(A, kmax, Rcpp=TRUE),
              partialColSortDesc(A, kmax, Rcpp=FALSE),
              replications=1)
}

#> Loading required package: rbenchmark

#>                                        test replications elapsed relative
#> 2 partialColSortDesc(A, kmax, Rcpp = FALSE)            1   1.115   10.421
#> 1  partialColSortDesc(A, kmax, Rcpp = TRUE)            1   0.107    1.000
#>   user.self sys.self user.child sys.child
#> 2     1.037    0.072          0         0
#> 1     0.092    0.014          0         0