post hoc bound obtained from Simes' inequality

Lower bound on the number of correct rejections using Simes' reference family

posthocBySimesRcpp(p, select, alpha)

posthocBySimes(p, select, alpha, verbose = FALSE)

Arguments

p
A numeric vector of m p-values for all tested hypotheses.

select
A vector of indices in \([1, \dots m]\) of the hypotheses to be selected.

alpha
A numeric value, the significance level of the test procedure.

verbose
If TRUE, prints verbose result to the screen. Defaults to FALSE.

Value

A integer value, Simes's lower bound on the number of correct rejections within the selected hypotheses

Details

If (R_k)_k provides jFWER control at level \(\alpha\) then with probability greater than \(1-\alpha\), \(|H_0 cap R| \leq \min_k {|R \cap (R_k)^c|+k-1}\) A bit better: \(|H_0 cap R| \leq (\min_{k<= |R|} {|R \cap R_k^c|+k-1}) \wedge |R|\)

Functions

posthocBySimesRcpp: Rcpp version

posthocBySimes: R version

Examples

m <- 1e3
m1 <- 200
p <- 1-pnorm(c(rnorm(m1, mean=4), rnorm(m-m1, mean=0)))
R <- sample(m, 10)
if (require("cherry")) {
  hom <- hommelFast(p)
  pickSimes(hom, R, silent=TRUE)
}

#> Loading required package: cherry

#> Loading required package: bitops

#> Loading required package: lpSolve

#> Loading required package: Matrix

#> Warning: package ‘Matrix’ was built under R version 3.4.2

#> Loading required package: slam

#> [1] 1

alpha <- 0.10
posthocBySimes(p, R, alpha=alpha)

#> [1] 1

posthocBySimesRcpp(p, R, alpha=alpha)

#> [1] 1