Prune a forest structure to speed up computations

The pruned forest structure makes the computation of V.star(), curve.V.star.forest.naive() and curve.V.star.forest.fast() faster.

Usage

pruning(C, ZL, leaf_list, prune.leafs = FALSE, delete.gaps = FALSE)

Arguments

C

A list of list representing the forest structure. See V.star() for more information.

ZL

A list of integer vectors representing the upper bounds \(\zeta_k\) of the forest structure. See V.star() for more information.

leaf_list

A list of vectors representing the atoms of the forest structure. See V.star() for more information.

prune.leafs

A boolean, FALSE by default. If TRUE, will also prune atoms/leafs for which \(\zeta_k \geq |R_k|\), this makes the computation of V.star() and curve.V.star.forest.naive() even faster but should not be used with curve.V.star.forest.fast() because this needs the structure to be complete (i.e., with all its atoms). This is why the default option is FALSE.

delete.gaps

A boolean, FALSE by default. If TRUE, will also delete the gaps in the structure induced by the pruning, see delete.gaps().

Value

A list with three named elements.

VstarNm

\(V^*(\mathbb N_m)\) is computed as by-product by the algorithm, so we might as well return it.

C

The new C after pruning.

ZL

The new ZL after pruning.

References

Durand, G., Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc false positive control for structured hypotheses. Scandinavian Journal of Statistics, 47(4), 1114-1148.

Durand G. (2025). A fast algorithm to compute a curve of confidence upper bounds for the False Discovery Proportion using a reference family with a forest structure. arXiv:2502.03849.