Welch t-tests for rows of a matrix

Welch t-tests for each row of a matrix, intended to be speed efficient

rowWelchTests(mat, categ = colnames(mat))

Arguments

mat
A numeric matrix whose rows correspond to variables and columns to observations

categ
A vector of ncol(mat) categories for the observations

Value

A list with class "htest" containing the following components:

statistic

the value of the t-statistics

parameter

the degrees of freedom for the t-statistics

p.value

the p-values for the tests

Details

Other functions to perform t-tests on each row of a matrix include: fastT: this function only calculates the test statistics and not the corresponding degrees of freedom or p-value; rowttests: this function calculates the test statistics, degrees of freedom and associated p-value, but only in the case of equal group variances (standard t-test and not Welch t-test)

References

B. L. Welch (1951), On the comparison of several mean values: an alternative approach. Biometrika, *38*, 330-336

Examples

p <- 1e3
n <- 38
mat <- matrix(rnorm(p*n), ncol=n)
cls <- rep(c(0, 1), times=c(27, n-27))
fwt <- rowWelchTests(mat, categ=cls)
str(fwt)

#> List of 3
#>  $ statistic: num [1:1000] 1.24 -1.851 0.802 -0.836 -0.501 ...
#>  $ parameter: num [1:1000] 17.5 20.1 24.9 15.9 20.1 ...
#>  $ p.value  : num [1:1000] 0.231 0.079 0.43 0.416 0.622 ...
#>  - attr(*, "class")= chr "htest"

# compare with ordinary t.test:
pwt <- apply(mat, 1, FUN=function(x) {
   t.test(x[cls==1], x[cls==0])$p.value
})
sum(abs(fwt$p.value-pwt))  ## same results

#> [1] 8.579117e-14