Volcano plot

Volcano plot for differential expression studies

Usage

# S3 method for class 'numeric'
volcanoPlot(
  x,
  y,
  pval,
  thr,
  p = 1,
  q = 1,
  r = 0,
  cex = c(0.2, 0.6),
  col = c("#33333333", "#FF0000", "#FF666633"),
  pch = 19,
  ylim = NULL,
  bounds = TRUE,
  ...
)

Arguments

x

A vector of fold changes (x axis of the volcano plot)

y

A vector of p-values (y axis of the volcano plot)

pval

A vector of p-values, of the same length as x, use to estimate post-hoc bounds

thr

A numeric vector of length K, a JER controlling family, used to estimate post-hoc bounds

p

A numeric value, the p-value threshold under which genes are selected

q

A numeric value, the q-value (or FDR-adjusted p-value) threshold under which genes are selected

r

A numeric value, the absolute fold change above which genes are selected

cex

A numeric vector of length 2, the relative magnification factor for unselected (cex[1]) and unselected (cex[2]) genes.

col

A vector of length 3

pch

An integer or single character string specifying the plotting character, see par

ylim

A numeric vector of length 2, the \(y\) limits of the plot

bounds

A boolean value: should the post hoc bounds be displayed on the plot? Defaults to TRUE

...

Not used

Value

The indices of selected genes (returned invisibly)

Details

A Welch T-test of differential expression between the two categories defined by categ are applied for each gene using the rowWelchTests function, which also outputs the "fold change" (mean difference in log scale) between the two categories.

See also