Create a complete dyadic tree structure

Create a complete dyadic tree structure

Create a complete dyadic tree structure

dyadic.from.leaf_list(leaf_list, method)

dyadic.from.window.size(m, s, method)

dyadic.from.height(m, H = NULL, method)

Arguments

leaf_list

A list of leaves

method

A numeric value. If method == 1, start from the leaves and group nodes of a same height 2 by 2 as long as possible. If method==2, start from the root and divide nodes in 2 nodes of equal size as long as possible

m

An integer value, the number of elements in the structure

s

An integer value, the number of elements in each leaf

H

An integer value, the desired maximal height of the tree

Value

A list of lists containing the dyadic structure

A list with two elements:

leaf_list

A list of leaves

C

A list of lists containing the dyadic structure

Details

dyadic.from.window.size

Dyadic tree structure from window size: the number of elements in each leaf is set to s

dyadic.from.height

Dyadic tree structure from height: the total height of tree is set to H

dyadic.from.max.size

Dyadic tree structure from maximum achievable height: the total height of tree is set to the maximum possible height, ie floor(2 + log2(m - 1))

Examples

m <- 6
dd <- dyadic.from.window.size(m, s = 2, method = 2)
str(dd)
#> List of 2
#>  $ leaf_list:List of 3
#>   ..$ : num [1:2] 1 2
#>   ..$ : num [1:2] 3 4
#>   ..$ : num [1:2] 5 6
#>  $ C        :List of 3
#>   ..$ :List of 1
#>   .. ..$ : num [1:2] 1 3
#>   ..$ :List of 2
#>   .. ..$ : num [1:2] 1 2
#>   .. ..$ : num [1:2] 3 3
#>   ..$ :List of 3
#>   .. ..$ : num [1:2] 1 1
#>   .. ..$ : num [1:2] 2 2
#>   .. ..$ : num [1:2] 3 3

dd <- dyadic.from.height(m, H = 3, method = 2)
str(dd)
#> List of 2
#>  $ leaf_list:List of 4
#>   ..$ : num [1:2] 1 2
#>   ..$ : num [1:2] 3 4
#>   ..$ : num 5
#>   ..$ : num 6
#>  $ C        :List of 3
#>   ..$ :List of 1
#>   .. ..$ : num [1:2] 1 4
#>   ..$ :List of 2
#>   .. ..$ : num [1:2] 1 2
#>   .. ..$ : num [1:2] 3 4
#>   ..$ :List of 4
#>   .. ..$ : num [1:2] 1 1
#>   .. ..$ : num [1:2] 2 2
#>   .. ..$ : num [1:2] 3 3
#>   .. ..$ : num [1:2] 4 4

dd <- dyadic.from.height(m, method=2)
str(dd)
#> List of 2
#>  $ leaf_list:List of 6
#>   ..$ : int 1
#>   ..$ : int 2
#>   ..$ : int 3
#>   ..$ : int 4
#>   ..$ : int 5
#>   ..$ : int 6
#>  $ C        :List of 4
#>   ..$ :List of 1
#>   .. ..$ : num [1:2] 1 6
#>   ..$ :List of 2
#>   .. ..$ : num [1:2] 1 3
#>   .. ..$ : num [1:2] 4 6
#>   ..$ :List of 4
#>   .. ..$ : num [1:2] 1 2
#>   .. ..$ : num [1:2] 3 3
#>   .. ..$ : num [1:2] 4 5
#>   .. ..$ : num [1:2] 6 6
#>   ..$ :List of 6
#>   .. ..$ : num [1:2] 1 1
#>   .. ..$ : num [1:2] 2 2
#>   .. ..$ : num [1:2] 3 3
#>   .. ..$ : num [1:2] 4 4
#>   .. ..$ : num [1:2] 5 5
#>   .. ..$ : num [1:2] 6 6

leaf_list <- dd$leaf_list