post hoc bound obtained from Simes' inequality

Lower bound on the number of correct rejections using Simes' reference family

posthocBySimes0Rcpp(p, select, alpha)

posthocBySimes(p, select, alpha, Rcpp = FALSE, verbose = FALSE)

Arguments

p
A numeric vector of m p-values for all tested hypotheses.

select
A vector of indices in \([1, \dots m]\) of the hypotheses to be selected.

alpha
A numeric value, the significance level of the test procedure.

Rcpp
A boolean value: use Rcpp version of the implementation? Defaults to FALSE.

verbose
If TRUE, prints verbose result to the screen. Defaults to FALSE.

Value

A integer value, Simes's lower bound on the number of correct rejections within the selected hypotheses

Details

If (R_k)_k provides jFWER control at level \(\alpha\) then with probability greater than \(1-\alpha\), \(|H_0 cap R| \leq \min_k {|R \cap (R_k)^c|+k-1}\) A bit better: \(|H_0 cap R| \leq (\min_{k<= |R|} {|R \cap R_k^c|+k-1}) \wedge |R|\)

Functions

posthocBySimes0Rcpp: Rcpp version

posthocBySimes: R version

Author

Gilles Blanchard, Pierre Neuvial and Etienne Roquain

Examples

m <- 1e3
m1 <- 200
p <- 1-pnorm(c(rnorm(m1, mean=4), rnorm(m-m1, mean=0)))
R <- union(1:10, sample(m, 10))
alpha <- 0.10
if (require("cherry")) {
  hom <- hommelFast(p)
  pickSimes(hom, R, silent=TRUE, alpha = alpha)
}

#> Loading required package: cherry

#> [1] 9

posthocBySimes(p, R, alpha=alpha)

#> [1] 9