Welch t-tests for rows of a matrix

Welch t-tests for each row of a matrix, intended to be speed efficient

rowWelchTests(mat, categ, alternative = c("two.sided", "less", "greater"))

Arguments

mat
A numeric matrix whose rows correspond to variables and columns to observations

categ
A numeric vector of ncol(mat) categories in \(0, 1\) for the observations

alternative
A character string specifying the alternative hypothesis. Must be one of "two.sided" (default), "greater" or "less". As in t.test, alternative = "greater" is the alternative that class 1 has a larger mean than class 0.

Value

A data.frame with containing the following columns:

statistic

the value of the t-statistics

parameter

the degrees of freedom for the t-statistics

p.value

the p-values for the tests

meanDiff

the mean difference

Details

Other functions to perform t-tests on each row of a matrix include: fastT: this function only calculates the test statistics and not the corresponding degrees of freedom or p-value; rowttests: this function calculates the test statistics, degrees of freedom and associated p-value, but only in the case of equal group variances (standard t-test and not Welch t-test)

References

B. L. Welch (1951), On the comparison of several mean values: an alternative approach. Biometrika, *38*, 330-336

Author

Pierre Neuvial

Examples

p <- 1e3
n <- 38
mat <- matrix(rnorm(p*n), ncol=n)
cls <- rep(c(0, 1), times=c(27, n-27))
fwt <- rowWelchTests(mat, cls, alternative = "greater")
str(fwt)

#> List of 4
#>  $ statistic: num [1:1000] 1.1429 -1.6283 -0.0639 0.4768 -0.9417 ...
#>  $ parameter: num [1:1000] 16.6 33 25.2 17.9 16.2 ...
#>  $ p.value  : num [1:1000] 0.135 0.944 0.525 0.32 0.82 ...
#>  $ meanDiff : num [1:1000] 0.3664 -0.4551 -0.0194 0.1886 -0.3804 ...

# compare with ordinary t.test:
pwt <- apply(mat, 1, FUN=function(x) {
   t.test(x[cls==1], x[cls==0], alternative = "greater")$p.value
})
all(abs(fwt$p.value-pwt) < 1e-10)  ## same results

#> [1] TRUE