Wilcoxon rank sum tests for each row of a matrix

Vectorized version of two-sample Wilcoxon rank sum tests

rowWilcoxonTests(
  mat,
  categ,
  alternative = c("two.sided", "less", "greater"),
  correct = TRUE
)

Arguments

mat
A numeric matrix whose rows correspond to variables and columns to observations

categ
A numeric vector of ncol(mat) categories in \(0, 1\) for the observations

alternative
A character string specifying the alternative hypothesis. Must be one of "two.sided" (default), "greater" or "less". As in wilcox.test, alternative = "greater" is the alternative that class 1 is shifted to the right of class 0.

correct
A logical indicating whether to apply continuity correction in the normal approximation for the p-value.

Value

A list with class "htest" containing the following components:

statistic

the value of the statistics

p.value

the p-values for the tests

A data.frame with columns

stat

A vector of m Wilcoxon sum rank test statistics of association between X and y.

stat0Mat

An m x B matrix of B realizations of a m-dimensional vector of test statistics under the null hypothesis of no association between X and y.

Details

The p-values are computed using the normal approximation as described in the wilcox.test function. The exact p-values (which can be useful for small samples with no ties) are not implemented (yet).

See also

Author

Gilles Blanchard, Pierre Neuvial and Etienne Roquain

Examples

p <- 100
n <- 120
mat <- matrix(rnorm(p*n), ncol = n)
cls <- rep(c(0, 1), times = c(n/2, n/2))
fwt <- rowWilcoxonTests(mat, categ = cls, alternative = "two.sided")
str(fwt)

#> 'data.frame':	100 obs. of  2 variables:
#>  $ statistic: num  1749 1762 2164 2068 1593 ...
#>  $ p.value  : num  0.791 0.844 0.0564 0.1603 0.2784 ...

# compare with ordinary wilcox.test:
pwt <- t(sapply(1:p, FUN=function(ii) {
  wt <- wilcox.test(mat[ii, cls==1], mat[ii, cls==0], alternative = "two.sided")
  c(statistic = wt[["statistic"]], p.value = wt[["p.value"]])
}))
all(abs(fwt$p.value-pwt[, "p.value"]) < 1e-10)  ## same results

#> [1] TRUE

all(abs(fwt$statistic-pwt[, "statistic.W"]) < 1e-10)  ## same results

#> [1] TRUE