Welch test from sufficient statistics

Welch test from sufficient statistics

suffWelchTest(
  mx,
  my,
  sx,
  sy,
  nx,
  ny,
  alternative = c("two.sided", "less", "greater")
)

Arguments

mx
A numeric value or vector, the sample average for condition "x"

my
A numeric value or vector of the same length as 'mx', the sample average for condition "y"

sx
A numeric value or vector of the same length as 'mx', the standard deviation for condition "x"

sy
A numeric value or vector of the same length as 'mx', the standard deviation for condition "y"

nx
A numeric value or vector of the same length as 'mx', the sample size for condition "x"

ny
A numeric value or vector of the same length as 'mx', the sample size for condition "y"

alternative
A character string specifying the alternative hypothesis. Currently only "two.sided" is implemented

Value

A list with elements:

statistic

the value of the t-statistic

parameter

the degrees of freedom for the t-statistic

p.value

the p-value for the test

Details

In accordance with the implementation of t.test and friends, we follow the rule that 'alternative = "greater"' is the alternative that 'x' has a larger mean than 'y'.

Examples

# Reproducing the results of the 't.test' function

x <- rnorm(100)
y <- rnorm(34, mean = 1)
target <- t.test(x, y)
swt <- suffWelchTest(mean(x), mean(y), sd(x), sd(y), length(x), length(y))
print(swt$statistic - target$statistic)

#> t 
#> 0 

print(swt$p.value - target$p.value)

#> [1] 3.882121e-17

print(swt$parameter-target$parameter)

#>            df 
#> -7.105427e-15 

target <- t.test(x, y, alternative = "greater")
swt <- suffWelchTest(mean(x), mean(y), sd(x), sd(y), length(x), length(y), alternative = "greater")
print(swt$statistic - target$statistic)

#> t 
#> 0 

print(swt$p.value - target$p.value)

#> [1] 0

print(swt$parameter-target$parameter)

#>            df 
#> -7.105427e-15