Estimate the number of true null hypotheses among a set of p-values

Estimate the number of true null hypotheses among a set of p-values

zeta.HB(pval, lambda)

zeta.trivial(pval, lambda)

zeta.DKWM(pval, lambda)

Arguments

pval

A vector of \(p\)-values

lambda

A numeric value in \([0,1]\), the target level of the test

Value

The numer of true nulls is estimated as follows:

zeta.DKWM

Dvoretzky-Kiefer-Wolfowitz-Massart inequality (related to the Storey estimator of the proportion of true nulls) with parameter lambda

zeta.HB

Holm-Bonferroni test with parameter lambda

zeta.trivial

the size of the p-value set (\(lambda\) is not used)

References

Durand, G., Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc false positive control for structured hypotheses. Scandinavian Journal of Statistics, 47(4), 1114-1148.

Dvoretzky, A., Kiefer, J., and Wolfowitz, J. (1956). Asymptotic minimax character of the sample distribution function and of the classical multinomial estimator. The Annals of Mathematical Statistics, pages 642-669.

Holm, S. A simple sequentially rejective multiple test procedure. Scandinavian Journal of Statistics 6 (1979), pp. 65-70.

Massart, P. (1990). The tight constant in the Dvoretzky-Kiefer-Wolfowitz inequality. The Annals of Probability, pages 1269-1283.

Storey, J. D. (2002). A direct approach to false discovery rates. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 64(3):479-498.

Examples

x <- rnorm(100, mean = c(rep(c(0, 2), each = 50)))
pval <- 1-pnorm(x)
zeta.trivial(pval)
#> [1] 100
zeta.HB(pval, 0.05)
#> [1] 90
zeta.DKWM(pval, 0.05)
#> [1] 57