Estimate of the proportion of true nulls in each node of a tree

Estimate of the proportion of true nulls in each node of a tree

zetas.tree(C, leaf_list, method, pvalues, alpha)

zetas.tree.refined(C, leaf_list, method, pvalues, alpha)

Arguments

C
Tree structure

leaf_list
A list of leaves

method
A function to compute the estimators

pvalues
A vector of \(p\)-values

alpha
A target level

Details

The proportion of true nulls in each node is estimated by an union bound on the regions. That is, the provided method is applied at level alpha/nR where nR is the number of regions.

In zetas.tree.refined, one tries to estimate the number of regions containing only signal.

References

Durand, G., Blanchard, G., Neuvial, P., & Roquain, E. (2020). Post hoc false positive control for structured hypotheses. Scandinavian Journal of Statistics, 47(4), 1114-1148.

Examples

m <- 1000 
dd <- dyadic.from.window.size(m, s = 10, method = 2)
leaf_list <- dd$leaf_list
mu <- gen.mu.leaves(m, K1 = 3, d = 1, grouped = FALSE, "const", barmu = 4, leaf_list)
pvalues<-gen.p.values(m, mu, rho = 0)
C <- dd$C 
ZL<-zetas.tree(C, leaf_list, zeta.DKWM, pvalues, alpha = 0.05)